+++ src/tr2/game/output.c
@@ -1093,8 +1093,8 @@ void Output_SwitchViewport(const VIEWPORT_SPACE space)
 {
     if (space == VIEWPORT_UI) {
         GFX_Renderer_BindUiFbo();
-        GFX_Context_Clear();
     }
     GFX_Context_SwitchToViewport(space);
     Render_SwitchViewport(space);
+    GFX_Context_Clear();
 }