+++ docs/tr1/CHANGELOG.md
@@ -36,6 +36,7 @@
 - fixed the collision box on some static meshes in Egypt to prevent the camera shaking when Lara walks by them (#762)
 - fixed incorrect wet footstep sounds in some of Lara's climb-up animations (#3607, regression from 4.6)
 - fixed the `/kill` command potentially causing a crash if used in a level with pods that don't hatch creatures (#3628, regression from 4.12)
+- fixed wireframe mode rendering as mostly white (#3649, regression from 4.13.2)
 - improved object loading error messages when an invalid object ID is detected
 - improved frames in Lara's jump-twist animations
 - removed the option in Unfinished Business to fix animated sprites as it is irrelevant there